reorganize kafka backend files
